mail/update_emailerr_script/requirements.txt
mysql.connector
python-dotenv

